crypto_secretbox: make generic around underlying stream cipher
Extracts a generic `SecretBox<C>` type which is generic around the
underlying stream cipher, with the goal of supporting both XSalsa20 and
XChaCha20.

For now only an `XSalsa20Poly1305` type alias is provided.

This also removes the former `XSalsa20Poly1305::generate_nonce` inherent
method which is now provided by the `AeadCore` trait.