crypto_secretbox: add `chacha20` feature (#98)
Adds `chacha20` feature providing a `XChaCha20Poly1305` type alias.

Note that this construction of "XChaCha20Poly1305" is different from the
one implemented by the `chacha20poly1305` crate, which implements
constructions based on the IETF variant of ChaCha20Poly1305.

The flavor of `XChaCha20Poly1305` in this crate is based on pre-IETF
constructions and is only supported by libsodium.

This is an unfortunate state of affairs to have two incompatible
constructions with the same name. One of the original goals of
`draft-irtf-cfrg-xchacha` was to ensure that "XChaCha20" was always
based on the IETF variant of ChaCha20.

The `AEAD_XChaCha20_Poly1305` construction that I-D defines is only
compatible with the version implemented by the `chacha20poly1305` crate,
and is *INCOMPATIBLE* with the construction provided by this crate,
which lacks features like AAD and is therefore not a full AEAD mode.

Comments to this effect have been left on the `XChaCha20Poly1305` type
definition in this crate.