crypto_secretbox: use `Salsa20` and `ChaCha20Legacy` (#100)
Previously this was using the IETF variant of ChaCha20, but for
compatibility with e.g. libsodium we need to use the original DJB
variant.

This also changes to using the `Salsa20` (as opposed to `XSalsa20`) type
and using the `Kdf` trait to impl both `XSalsa20` and `XChaCha20`, since
the `chacha20` version of `XChaCha20` is built on the IETF variant of
`ChaCha20` (so as to be compatible with draft-irtf-cfrg-xchacha).