test-vector-gen: initial `crypto_secretbox` test vectors (#111)
Adds a `publish = false` crate for generating test vectors using
libsodium.

Uses the `libsodium-sys-stable` crate, which unlike `sodiumoxide` and
`libsodium-sys` is actually maintained (by libsodium's author, no less).
This allows the test vectors to be generated once so libsodium isn't a
hard dependency for development work.

Unfortunately both the `libsodium-sys` and `libsodium-sys-stable`
packages define a `libsodium_sys` crate, so it's not possible to use
both in the same project. For that reason this removes the
`dev-dependencies` on `sodiumoxide`, with TODOs to refactor the code
which previously used `sodiumoxide` into a test vector generator.

It prints out test vector strings for given inputs, using the syntax of
the `hex-literal` crate, i.e. `hex!`.

It currently reproduces the existing XSalsa20Poly1305 test vectors
(originally extracted from NaCl), and generates new test vectors for the
libsodium-flavored XChaCha20Poly1305.

These test vectors are used to test the existing `XChaCha20Poly1305`
implementation in `crypto_secretbox`, which passes. It was previously
untested (and also unreleased).